Add `--replace` to `update_data`.

+++ lib/sf_covid19/cli.rb
@@ -20,13 +20,18 @@ module SFCOVID19
     end
 
     desc 'update_data PATH', 'Update historic data at path in JSON format'
+    option :replace, type: :boolean, desc: 'Replace content of path'
     option :pretty, type: :boolean, desc: 'Make JSON pretty.'
     def update_data(path)
       historic_data = JSON.parse(File.read(path))
       new_data = SFDPHScraper.new.scrape!
       data = historic_data.merge(new_data)
       json = options[:pretty] ? JSON.pretty_generate(data) : JSON.generate(data)
-      File.open(path, 'w') { |file| file.puts json }
+      if options[:replace]
+        File.open(path, 'w') { |file| file.puts json }
+      else
+        puts json
+      end
     end
 
     TIMESTAMP_COLUMN_KEY = 'timestamp'
